@@ -13,7 +13,7 @@ public record Borders
         public static Shadows DefaultShadow = Shadows.Default;
 
         public static BorderRadius DefaultRoundedBorderRadius = BorderRadius.Default;
-        public static Shadows DefaultWithShadowBorderShadow = Shadows.Small;
+        public static Shadows DefaultWithShadow = Shadows.Small;
         
         public static BorderStyle DefaultHeaderStyle = BorderStyle.Bottom;
         public static BorderStyle DefaultFooterStyle = BorderStyle.Top;
@@ -23,9 +23,10 @@ public record Borders
     private static Borders? _defaultWithShadow;
     private static Borders? _defaultRounded;
     private static Borders? _defaultRoundedWithShadow;
-    private static Borders? _defaultHeader;
-    private static Borders? _defaultFooter;
-    
+    private static Borders? _defaultForHeaders;
+    private static Borders? _defaultForFooters;
+    private static Borders? _none;
+
     public static Borders Default => _default ??= new Borders
         {
             Style = Standards.DefaultStyle,
@@ -43,24 +44,30 @@ public record Borders
     
     public static Borders DefaultWithShadow => _defaultWithShadow ??= Default with
     {
-        Shadow = Standards.DefaultWithShadowBorderShadow
+        Shadow = Standards.DefaultWithShadow
     };
     
     public static Borders DefaultRoundedWithShadow => _defaultRoundedWithShadow ??= DefaultRounded with
     {
-        Radius = Standards.DefaultRoundedBorderRadius
+        Radius = Standards.DefaultRoundedBorderRadius,
+        Shadow = Standards.DefaultWithShadow
     };
 
-    public static Borders DefaultHeader => _defaultHeader ??= Default with
+    public static Borders DefaultForHeaders => _defaultForHeaders ??= Default with
     {
         Style = Standards.DefaultHeaderStyle
     };
     
-    public static Borders DefaultFooter => _defaultFooter ??= Default with
+    public static Borders DefaultForFooters => _defaultForFooters ??= Default with
     {
         Style = Standards.DefaultFooterStyle
     };
-    
+
+    public static Borders DefaultNone => _none ??= Default with
+    {
+        Style = BorderStyle.None
+    };
+
     private string? cssClasses;
     
     public BorderStyle Style { get; init; }
@@ -82,15 +89,15 @@ public record Borders
     
     private IEnumerable<string> AllCssClasses()
     {
-        yield return GetStyleCssClasses();
-        yield return GetColorCssClasses();
-        yield return GetWidthCssClasses();
-        yield return GetRadiusCssClasses();
-        yield return GetRoundedSizeCssClasses();
-        yield return Shadow.GetCssClass();
+        yield return ToStyleCssClasses();
+        yield return ToColorCssClasses();
+        yield return ToWidthCssClasses();
+        yield return ToRadiusCssClasses();
+        yield return ToRoundedSizeCssClasses();
+        yield return Shadow.ToCssClass();
     }
 
-    private string GetStyleCssClasses()
+    private string ToStyleCssClasses()
     {
         return Style switch
         {
@@ -107,7 +114,7 @@ public record Borders
         };
     }
 
-    private string GetColorCssClasses()
+    private string ToColorCssClasses()
     {
         return Color switch
         {
@@ -126,7 +133,7 @@ public record Borders
         };
     }
 
-    private string GetWidthCssClasses()
+    private string ToWidthCssClasses()
     {
         return Width switch
         {
@@ -139,7 +146,7 @@ public record Borders
         };
     }
 
-    private string GetRadiusCssClasses()
+    private string ToRadiusCssClasses()
     {
         return Radius switch
         {
@@ -155,7 +162,7 @@ public record Borders
         };
     }
 
-    private string GetRoundedSizeCssClasses()
+    private string ToRoundedSizeCssClasses()
     {
         return RoundedSize switch
         {
