@@ -3,13 +3,16 @@
 <AppLayout>
     <TopStart>
         <AppBrand></AppBrand>
-        <AppMenuButton>M</AppMenuButton>
+        <AppMenuButton>
+            <Icon Kind="BsIconNames.List" />
+        </AppMenuButton>
     </TopStart>
     <Menu>
         <AppNavMenu>
             <AppNavItem Label="Home"/>
             <AppNavItem Label="Counter" HRef="counter"/>
             <AppNavItem Label="Fetch Data" HRef="fetchdata"/>
+            <AppNavItem Label="Blazor Show" HRef="show" />
             <AppNavItem Label="Containers" HRef="containers"/>
             <AppNavItem Label="Spinners" HRef="spinners"/>
             <AppNavItem Label="App Layout" HRef="applayout"/>
