@@ -25,7 +25,7 @@
     public Borders PanelBorder { get; set; } = Borders.Default;
 
     [Parameter]
-    public Borders HeaderBorder { get; set; } = Borders.DefaultHeader;
+    public Borders HeaderBorder { get; set; } = Borders.DefaultForHeaders;
         
     [Parameter]
     public string? PanelAdditionalClasses { get; set; }
