@@ -24,10 +24,10 @@
     public Borders PanelBorder { get; set; } = Borders.Default;
 
     [Parameter]
-    public Borders HeaderBorder { get; set; } = Borders.DefaultHeader;
+    public Borders HeaderBorder { get; set; } = Borders.DefaultForHeaders;
     
     [Parameter]
-    public Borders FooterBorder { get; set; } = Borders.DefaultFooter;
+    public Borders FooterBorder { get; set; } = Borders.DefaultForFooters;
     
     [Parameter]
     public string? PanelAdditionalClasses { get; set; }
